# docs/storoom.api-docs.yaml
openapi: 3.0.1
info:
  title: storoom
  version: v1
servers:
- url: https://localhost:8080/
security:
- Bearer Authentication: []
paths:
  /users/{username}/type:
    put:
      tags:
      - Users
      description: Update userType by user's username.
      operationId: updateUserType
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUserTypeDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /users/{username}/password:
    put:
      tags:
      - Users
      description: Update password by user's username.
      operationId: updatePassword
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPasswordDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /units/{id}:
    put:
      tags:
      - Units
      description: Update unit info.
      operationId: updateUnitInfo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitInfoDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnitFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Units
      description: Delete unit.
      operationId: deleteUnit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnitFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /units/{id}/status:
    put:
      tags:
      - Units
      description: Update unit status.
      operationId: updateUnitStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitStatusDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnitFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
...
  /users:
    get:
      tags:
      - Users
      description: Get users
      operationId: getUsers
      parameters:
      - name: userType
        in: query
        description: (optional) userType to search for
        required: false
        schema:
          type: string
      - name: page
        description: number of page
      - name: size
        description: size of page
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUserReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
    post:
      tags:
      - Users
      description: Create user.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserFullDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "409":
          description: Request action conflicts with the current state of things.
  /units:
    get:
      tags:
      - Units
      description: Get all units.
      operationId: getUnits
      parameters:
      - name: page
        description: number of page
      - name: size
        description: size of page
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUnitDTO'
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
    post:
      tags:
      - Units
      description: Create unit.
      operationId: createUnit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnitFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "409":
          description: Request action conflicts with the current state of things.
...
  /users/{username}:
    delete:
      tags:
      - Users
      description: Delete user by username.
      operationId: deleteByUsername
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
components:
  schemas:
    UserUserTypeDTO:
      type: object
      properties:
        userType:
          type: string
          enum:
          - admin
          - superuser
          - client
    UserReadDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        userType:
          type: string
          enum:
          - admin
          - superuser
          - client
    UserPasswordDTO:
      type: object
      properties:
        password:
          type: string
    UnitInfoDTO:
      type: object
      properties:
        sizeX:
          type: integer
          format: int32
        sizeY:
          type: integer
          format: int32
        sizeZ:
          type: integer
          format: int32
        unitType:
          type: string
          enum:
          - S
          - M
          - L
          - XL
        locationId:
          type: string
          format: uuid
        lockId:
          type: string
          format: uuid
    LocationDTO:
      type: object
      properties:
        address:
          type: string
        locationType:
          type: string
          enum:
          - stand
          - warehouse
          - main_office
        ownerIds:
          type: array
          items:
            type: string
            format: uuid
    LockDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        manufacturer:
          type: string
          format: uuid
    UnitFullDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sizeX:
          type: integer
          format: int32
        sizeY:
          type: integer
          format: int32
        sizeZ:
          type: integer
          format: int32
        unitType:
          type: string
          enum:
          - S
          - M
          - L
          - XL
        status:
          type: string
          enum:
          - dead
          - available
          - occupied
          - pending
        location:
          $ref: '#/components/schemas/LocationDTO'
        lock:
          $ref: '#/components/schemas/LockDTO'
    UnitStatusDTO:
      type: object
      properties:
        status:
          type: string
          enum:
          - dead
          - available
          - occupied
          - pending
...
    UserFullDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        password:
          type: string
        userType:
          type: string
          enum:
          - admin
          - superuser
          - client
    UnitDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sizeX:
          type: integer
          format: int32
        sizeY:
          type: integer
          format: int32
        sizeZ:
          type: integer
          format: int32
        unitType:
          type: string
          enum:
          - S
          - M
          - L
          - XL
        status:
          type: string
          enum:
          - dead
          - available
          - occupied
          - pending
        locationId:
          type: string
          format: uuid
        lockId:
          type: string
          format: uuid
...
    PageUserReadDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserReadDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageUnitDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/UnitDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT
    Basic Authentication:
      type: http
      scheme: basic